+++ app/Widgets/ContactBlogConfig/contactblogconfig.js
@@ -1,0 +1,3 @@
+MovimWebsocket.attach(() => {
+    ContactBlogConfig_ajaxCheckAccessModel(MovimUtils.urlParts().params[0]);
+});