+++ src/Moxl/Xec/Action/Pubsub/GetConfig.php
@@ -25,8 +25,17 @@ class GetConfig extends Action
 
     public function handle(?\SimpleXMLElement $stanza = null, ?\SimpleXMLElement $parent = null)
     {
+        $value = $stanza->pubsub->configure->xpath('//field[@var=\'pubsub#access_model\']/value/text()');
+
+        $accessModel = null;
+
+        if (is_array($value) && count($value) > 0) {
+            $accessModel = (string)$value[0];
+        }
+
         $this->pack([
             'config' => $stanza->pubsub->configure,
+            'access_model' => $accessModel,
             'server' => $this->_to,
             'node' => $this->_node,
             'advanced' => $this->_advanced